implement content repos with override to INIT, RESET or UPDATE repository
create tests for INIT, UPDATE and RESET behavior

+++ src/test/groovy/com/cloudogu/gitops/utils/TestScmmRepoProvider.groovy
@@ -10,7 +10,7 @@ import static org.mockito.Mockito.spy
 
 class TestScmmRepoProvider extends ScmmRepoProvider {
     Map<String, ScmmRepo> repos = [:]
-    
+
     TestScmmRepoProvider(Config config, FileSystemUtils fileSystemUtils) {
         super(config, fileSystemUtils)
     }
@@ -23,27 +23,45 @@ class TestScmmRepoProvider extends ScmmRepoProvider {
         }
 
         ScmmRepo repo = new ScmmRepo(config, repoTarget, fileSystemUtils) {
+
+            String remoteGitRepopUrl = ''
+
             @Override
-            protected String getGitRepositoryUrl() {
-                def tempDir = File.createTempDir('gitops-playground-repocopy')
-                tempDir.deleteOnExit()
-                def originalRepo = System.getProperty("user.dir") + "/src/test/groovy/com/cloudogu/gitops/utils/data/git-repository/"
+            String getGitRepositoryUrl() {
+                if (!remoteGitRepopUrl) {
 
-                FileUtils.copyDirectory(new File(originalRepo), tempDir)
+                    def tempDir = File.createTempDir('gitops-playground-repocopy')
+                    tempDir.deleteOnExit()
+                    def originalRepo = System.getProperty("user.dir") + "/src/test/groovy/com/cloudogu/gitops/utils/data/git-repository/"
 
-                return 'file://' + tempDir.absolutePath
+                    FileUtils.copyDirectory(new File(originalRepo), tempDir)
+                    remoteGitRepopUrl = 'file://' + tempDir.absolutePath
+                }
+                return remoteGitRepopUrl
+
             }
 
             @Override
             protected Git gitClone() {
                 // Cloning from filepath does not work without setting branch
-                Git.cloneRepository()
-                        .setURI(getGitRepositoryUrl())
-                        .setDirectory(new File(absoluteLocalRepoTmpDir))
-                        .setNoCheckout(true)
-                        .setBranch('main')
-                        .call()
+                try {
+                    Git.cloneRepository()
+                            .setURI(getGitRepositoryUrl())
+                            .setDirectory(new File(absoluteLocalRepoTmpDir))
+                            .setNoCheckout(true)
+                            .setBranch('main')
+                            .call()
+
+                } catch (Exception e) {
+                    // test workaround for testing same repo again. Clean folder with .git and do it again.
+                    // it need 2-3 tries
+                    fileSystemUtils.deleteFilesExcept(new File(absoluteLocalRepoTmpDir))
+                    gitClone()
+                }
+
+
             }
+
         }
         // Create a spy to enable verification while keeping real behavior
         ScmmRepo spyRepo = spy(repo)
